Don't raise from check() under normal operation.
check() checks whether a domain is available or not. If a domain is
available, the API returns a 404 error code, which causes an error
to be raised. Having to catch an error for normal operation is awkward,
so this change supresses the error caused by the 404 HTTP code.

dnsimple docs for check():
http://developer.dnsimple.com/domains/#check-domain-availability